Move needs-rebase external plugin into 'prow/external-plugins/'.
