+++ prow/external-plugins/needs-rebase/BUILD
@@ -3,10 +3,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["main.go"],
-    importpath = "k8s.io/test-infra/experiment/needs-rebase",
+    importpath = "k8s.io/test-infra/prow/external-plugins/needs-rebase",
     visibility = ["//visibility:private"],
     deps = [
-        "//experiment/needs-rebase/plugin:go_default_library",
+        "//prow/external-plugins/needs-rebase/plugin:go_default_library",
         "//prow/github:go_default_library",
         "//prow/hook:go_default_library",
         "//prow/pluginhelp/externalplugins:go_default_library",
@@ -18,7 +18,7 @@ go_library(
 go_binary(
     name = "needs-rebase",
     embed = [":go_default_library"],
-    importpath = "k8s.io/test-infra/experiment/needs-rebase",
+    importpath = "k8s.io/test-infra/prow/external-plugins/needs-rebase",
     visibility = ["//visibility:public"],
 )
 
@@ -33,7 +33,7 @@ filegroup(
     name = "all-srcs",
     srcs = [
         ":package-srcs",
-        "//experiment/needs-rebase/plugin:all-srcs",
+        "//prow/external-plugins/needs-rebase/plugin:all-srcs",
     ],
     tags = ["automanaged"],
     visibility = ["//visibility:public"],
