+++ prow/external-plugins/needs-rebase/plugin/BUILD
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = ["plugin.go"],
-    importpath = "k8s.io/test-infra/experiment/needs-rebase/plugin",
+    importpath = "k8s.io/test-infra/prow/external-plugins/needs-rebase/plugin",
     visibility = ["//visibility:public"],
     deps = [
         "//prow/github:go_default_library",
@@ -18,7 +18,7 @@ go_test(
     name = "go_default_test",
     srcs = ["plugin_test.go"],
     embed = [":go_default_library"],
-    importpath = "k8s.io/test-infra/experiment/needs-rebase/plugin",
+    importpath = "k8s.io/test-infra/prow/external-plugins/needs-rebase/plugin",
     deps = [
         "//prow/github:go_default_library",
         "//prow/plugins:go_default_library",
